renamed stack of stacks to set of stacks
