+++ README.md
@@ -27,9 +27,9 @@ This project is Licensed under the [MIT License](/LICENSE.md).
 * Queue ([JS](/data-structures/queue.js))
 * Queue with Two Stacks (CtCI)([JS](/data-structures/queue-with-two-stacks.js))
 * Singly Linked List ([R](/data-structures/singly_linked_list.rb))
+* Set of Stacks (CtCI)([R](/data-structures/set_of_stacks.rb))
 * Stack ([JS](/data-structures/stack.js))([R](/data-structures/linked_stack.rb))
 * Stack Min (CtCI)([R](/data-structures/stack_min.rb))
-* Stack of Stacks (CtCI)([R](/data-structures/stack_of_stacks.rb))
 * Stack with Two Queues (CtCI)([JS](/data-structures/stack-with-two-queues.js))
 * Tree ([JS](/data-structures/tree.js))
 * Trie ([JS](/data-structures/trie.js))
