+++ data-structures/set_of_stacks.rb
